requirements.txt
fastmcp
psycopg[binary]
python-dotenv
